+++ src/main/java/maeilmail/bulksend/policy/SequenceChoicePolicy.java
@@ -1,13 +1,13 @@
-package maeilmail.subscribequestion.policy;
+package maeilmail.bulksend.policy;
 
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import maeilmail.bulksend.ChoiceQuestionPolicy;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
 import maeilmail.subscribe.command.domain.Subscribe;
-import maeilmail.subscribequestion.ChoiceQuestionPolicy;
 import org.springframework.stereotype.Component;
 
 @Component
