+++ src/main/java/maeilmail/bulksend/policy/WeekdaysPersonalSequenceChoicePolicy.java
@@ -1,12 +1,12 @@
-package maeilmail.subscribequestion.policy;
+package maeilmail.bulksend.policy;
 
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import maeilmail.bulksend.ChoiceQuestionPolicy;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
 import maeilmail.subscribe.command.domain.Subscribe;
-import maeilmail.subscribequestion.ChoiceQuestionPolicy;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
