+++ src/test/java/maeilmail/bulksend/SendQuestionSchedulerTest.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribequestion;
+package maeilmail.bulksend;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
